Fix videoSource capture frame bug #1
There is a sporadic concurrency issue, between the gStreamer writing to the RingBuffer, and a client consuming frames from the RingBuffer. The issue is super hard to reproduce but what happens is that during a client reading a frame from the RingBuffer a new frame comes from the gStreamer on the other "Producer" thread, and our consumer happens to read this new Frame. However then a new consumption of the next frame comes, which goes through the mWaitEvent since it's unblocked by the second gStreamer frame, but when it tries to read the latest frame from the RingBuffer it fails, because it was already consumed by the first Capture call. This then causes the Capture to fail, because it uses the ReadLatestOnce flag.

This PR attempts to fix this concurrency issue by resetting the mWaitEvent the moment a new frame is reported by the gStreamer, so that the second Consume can wait. The issue was still reproduced but it does not happen as ofter now, which should mean that the fix is actually working.

Additionally this PR also adds more [ameru] prefixed logs, and timestamps to the logs, so that we can further troubleshoot this issue.